refactor: move startup logic to per-indexer container (#327)
